Update Sample_DigitalTwin: separate connection codes to another script
