+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_DigitalTwin/DigitalTwinConnector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1844a7138b00a3b48b32d56158ec4557
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
