+++ toio-sdk-unity/Assets/toio-sdk/Samples/Sample_DigitalTwin/Sample_DigitalTwin.unity
@@ -165,7 +165,7 @@ PrefabInstance:
     - target: {fileID: 8138842908862602432, guid: cdb3cb207738d3b45ade7abf174b8f7a,
         type: 3}
       propertyPath: m_RootOrder
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 8138842908862602432, guid: cdb3cb207738d3b45ade7abf174b8f7a,
         type: 3}
@@ -255,7 +255,7 @@ PrefabInstance:
     - target: {fileID: 5597929066422657874, guid: 8bd6ea65c87536d49ac5878fd9e412fe,
         type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 5597929066422657874, guid: 8bd6ea65c87536d49ac5878fd9e412fe,
         type: 3}
@@ -324,6 +324,74 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 11db91d6fc205dedaa64396bef9937c2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1 &815486382
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 815486385}
+  - component: {fileID: 815486384}
+  - component: {fileID: 815486383}
+  m_Layer: 0
+  m_Name: Connector
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &815486383
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 815486382}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1844a7138b00a3b48b32d56158ec4557, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  connectType: 2
+  localNamesToConnect:
+  - toio-t5E
+  - Cube
+  - Cube2
+  logScanned: 1
+--- !u!114 &815486384
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 815486382}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2097a692105d9498294c5417b183c53d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  simulators:
+  - {fileID: 678647617}
+  - {fileID: 1208939862}
+  mat: {fileID: 295603513}
+  mappingMethod: 1
+--- !u!4 &815486385
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 815486382}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &840845203
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -340,7 +408,7 @@ PrefabInstance:
     - target: {fileID: 5597929066422657874, guid: 8bd6ea65c87536d49ac5878fd9e412fe,
         type: 3}
       propertyPath: m_RootOrder
-      value: 3
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 5597929066422657874, guid: 8bd6ea65c87536d49ac5878fd9e412fe,
         type: 3}
@@ -407,7 +475,6 @@ GameObject:
   m_Component:
   - component: {fileID: 1030505189}
   - component: {fileID: 1030505190}
-  - component: {fileID: 1030505191}
   m_Layer: 0
   m_Name: Main Script
   m_TagString: Untagged
@@ -428,7 +495,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 0
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1030505190
 MonoBehaviour:
@@ -442,28 +509,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 865bb5287c78f4874ab940f0a04d11ae, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  connectType: 2
-  localNamesToConnect:
-  - toio-t5E
-  - Cube
-  - Cube2
---- !u!114 &1030505191
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1030505187}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2097a692105d9498294c5417b183c53d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  simulators:
-  - {fileID: 678647617}
-  - {fileID: 1208939862}
-  mat: {fileID: 295603513}
-  mappingMethod: 1
+  connector: {fileID: 815486383}
 --- !u!43 &1154838316
 Mesh:
   m_ObjectHideFlags: 0
